% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adis16405}
\alias{adis16405}
\title{Six samples collected from ADIS 16405 IMU at 100 Hz  over periods of about 3 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
adis16405
}
\description{
This data set contains six samples collected from ADIS 16405 IMU at 100 Hz.
}
\examples{
\dontrun{
data(adis16405)
adis16405
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
