% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADIS16405imu100Hz}
\alias{ADIS16405imu100Hz}
\title{Six samples collected from ADIS 16405 IMU at 100 Hz  over periods of about 3 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
ADIS16405imu100Hz
}
\description{
This data set contains six samples collected from ADIS 16405 IMU at 100 Hz.
}
\examples{
\dontrun{
data(ADIS16405imu100Hz)
ADIS16405imu100Hz
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
