% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adis_16405_imu5}
\alias{adis_16405_imu5}
\title{IMU Data from an ADIS 16405 IMU at 100 Hz  over periods of about 3 hours}
\format{An \code{imu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
adis_16405_imu5
}
\description{
This data set contains gyroscope and accelerometer data from ADIS 16405 IMU at 100 Hz.
}
\examples{
\dontrun{
data(adis_16405_imu5)
adis_16405_imu5
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
