% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cont_imu1.R
\docType{data}
\name{cont_imu1}
\alias{cont_imu1}
\title{Containment IMU Data from a XSens MTi-G sensor}
\format{A \code{matrix} with 500,000 observations and 1 variable. The variable is as follows:
\itemize{
 \item{Gyro. X}{Gyroscope X-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
cont_imu1
}
\description{
This data set contains gyroscope data
}
\examples{
\dontrun{
data(cont_imu1)
summary(cont_imu1)
}
}
\author{
JJB
}

