% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gmwmdata.R
\docType{package}
\name{gmwmdata}
\alias{gmwmdata}
\alias{gmwmdata-package}
\title{GMWM Data Package}
\description{
This package is meant to serve as a data specific example package that can be used with the GMWM
to model latent processes within a time series.
}
\details{
\tabular{ll}{
Package: \tab gmwmdata\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-09-16\cr
License: \tab Q + file LICENSE\cr
}
}
\examples{
# Please see the datasets
}
\author{
James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}

Stephane Guerrier \email{stephane@illinois.edu}
}

