% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imar.gyro}
\alias{imar.gyro}
\title{IMU Data over 4 Hours from IMAR Gyroscopes}
\format{An \code{imu} with 5,760,000 observations and 3 sensors. The sensors are defined as follows:
\describe{
 \item{\code{Gyro. X}}{Gyroscope X-Axis}
 \item{\code{Gyro. Y}}{Gyroscope Y-Axis}
 \item{\code{Gyro. Z}}{Gyroscope Z-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
imar.gyro
}
\description{
This data set contains gyroscope from multiple IMAR Gyroscopes.
}
\examples{
\dontrun{
data(imar.gyro)
imar.gyro
}
}
\author{
JJB
}
\keyword{datasets}

