% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imudata.R
\docType{package}
\name{imudata}
\alias{imudata}
\alias{imudata-package}
\title{IMU Data Package for the GMWM Methods Package in R}
\description{
This package is meant to serve as a data collection for IMU data. 
The data found here can be used with the GMWM methods R package to model 
latent processes within a time series.
}
\details{
\tabular{ll}{
Package: \tab gmwmdata\cr
Type: \tab Package\cr
Version: \tab 3.0.0\cr
Date: \tab 2016-03-05\cr
License: \tab CC BY-NC-SA 4.0\cr
}
}
\examples{
# Please see the datasets
}
\seealso{
\code{\link[imudata]{cont.imu1}}, \code{\link[imudata]{imu6}}, \code{\link[imudata]{imar.gyro}}, \code{\link[imudata]{ln200.accel}}, \code{\link[imudata]{ln200.gyro}}, \code{\link[imudata]{mtig}}, \code{\link[imudata]{navchip}} and \code{\link[imudata]{kvh}}
}
\author{
James Balamuta \email{balamut2@illinois.edu},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}

James Balamuta \email{balamut2@illinois.edu}
}
