% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvh}
\alias{kvh}
\title{Six samples collected from KVH1750 IMU (only accelerometers) over periods of about 3 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
kvh
}
\description{
This data set contains six samples collected from KVH1750 IMU.
}
\examples{
\dontrun{
data(kvh)
kvh
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
