% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvh1750.gyro}
\alias{kvh1750.gyro}
\title{Six samples collected from KVH1750 IMU at 100Hrz (only accelerometers) over periods of about 3 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
kvh1750.gyro
}
\description{
This data set contains six samples collected from KVH1750 IMU.
}
\examples{
\dontrun{
data(kvh1750.gyro)
kvh1750.gyro
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
