% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KVH1750imu1kHzGyro}
\alias{KVH1750imu1kHzGyro}
\title{Six samples collected from KVH1750 IMU at 1000 Hz (only gyroscopes) over periods of about 10 minutes}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
KVH1750imu1kHzGyro
}
\description{
This data set contains six samples collected from KVH1750 IMU at 1000 Hz.
}
\examples{
\dontrun{
data(KVH1750imu1kHzGyro)
KVH1750imu1kHzGyro
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
