% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{KVH1750imuAcc}
\alias{KVH1750imuAcc}
\title{Six samples collected from KVH1750 IMU (only accelerometers) over periods of about 3 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
KVH1750imuAcc
}
\description{
This data set contains six samples collected from KVH1750 IMU.
}
\examples{
\dontrun{
data(KVH1750imu100HzAcc)
KVH1750imu100HzAcc
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
