% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kvh_100hz_imu4}
\alias{kvh_100hz_imu4}
\title{IMU Data from a KVH1750 IMU at 100hrz over periods of about 3 hours.}
\format{An \code{imu} with over 1,000,000 observations and 6 sensors. The sensors are defined as follows:
\describe{
 \item{\code{Gyro. X}}{Gyroscope X-Axis}
 \item{\code{Gyro. Y}}{Gyroscope Y-Axis}
 \item{\code{Gyro. Z}}{Gyroscope Z-Axis}
 \item{\code{Accel. X}}{Accelerometer X-Axis}
 \item{\code{Accel. Y}}{Accelerometer Y-Axis}
 \item{\code{Accel. Z}}{Accelerometer Z-Axis}
}}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
kvh_100hz_imu4
}
\description{
This data set contains gyroscope and accelerometer data from KVH1750 IMU.
}
\examples{
\dontrun{
data(kvh_100hz_imu4)
kvh_100hz_imu4
}
}
\keyword{datasets}
