% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ln200.accel}
\alias{ln200.accel}
\title{IMU Data over 6 Hours from LN200 Accelerometers}
\format{An \code{imu} with 8,640,000 observations and 3 sensors. The sensors are defined as follows:
\describe{
 \item{\code{Accel. X}}{Accelerometer X-Axis}
 \item{\code{Accel. Y}}{Accelerometer Y-Axis}
 \item{\code{Accel. Z}}{Accelerometer Z-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
ln200.accel
}
\description{
This data set contains gyroscope from multiple LN200 Accelerometers
}
\examples{
\dontrun{
data(ln200.accel)
ln200.accel
}
}
\author{
JJB
}
\keyword{datasets}
