% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ln200.gyro}
\alias{ln200.gyro}
\title{IMU Data over 6 Hours from LN200 Gyroscopes}
\format{An \code{imu} with 8,640,000 observations and 3 sensors. The sensors are defined as follows:
\describe{
 \item{\code{Gyro. X}}{Gyroscope X-Axis}
 \item{\code{Gyro. Y}}{Gyroscope Y-Axis}
 \item{\code{Gyro. Z}}{Gyroscope Z-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
ln200.gyro
}
\description{
This data set contains gyroscope from multiple LN200 Gyroscopes.
}
\examples{
\dontrun{
data(ln200.gyro)
ln200.gyro
}
}
\author{
JJB
}
\keyword{datasets}

