% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtig}
\alias{mtig}
\title{IMU Data from a XSens MTi-G sensor}
\format{An \code{imu} with 873,684 observations and 6 sensors. The sensors are defined as follows:
\describe{
 \item{\code{Gyro. X}}{Gyroscope X-Axis}
 \item{\code{Gyro. Y}}{Gyroscope Y-Axis}
 \item{\code{Gyro. Z}}{Gyroscope Z-Axis}
 \item{\code{Accel. X}}{Accelerometer X-Axis}
 \item{\code{Accel. Y}}{Accelerometer Y-Axis}
 \item{\code{Accel. Z}}{Accelerometer Z-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
mtig
}
\description{
This data set contains gyroscope and accelerometer data from a XSens MTi-G sensor.
}
\details{
This data set is equivalent to the \code{\link[imudata]{imu6}} data set. The only difference
is that this data set acts as a way to showcase the results of the \code{\link[gmwm]{imu}} cast function whereas 
\code{\link[imudata]{imu6}} has yet to be cast using \code{\link[gmwm]{imu}}.
}
\examples{
\dontrun{
data(mtig)
mtig
}
}
\author{
JJB
}
\seealso{
\code{\link[imudata]{mtig}}
}
\keyword{datasets}

