% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtig100hrz}
\alias{mtig100hrz}
\title{Six samples collected from MTI-G-710 IMU at 100 Hz over periods of about 3 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
mtig100hrz
}
\description{
This data set contains six samples collected from MTI-G-710 IMU at 100 Hz.
}
\examples{
\dontrun{
data(mtig100hrz)
mtig100hrz
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
