% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtig1khrz}
\alias{mtig1khrz}
\title{Six samples collected from MTI-G-710 IMU at 1000 Hz over periods of about 10 minutes}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
mtig1khrz
}
\description{
This data set contains six samples collected from MTI-G-710 IMU at 1000 Hz.
}
\examples{
\dontrun{
data(mtig1khrz)
mtig1khrz
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
