% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTIG710imu1kHz}
\alias{MTIG710imu1kHz}
\title{Six samples collected from MTI-G-710 IMU at 1000 Hz (only accelerometers) over periods of about 10 minutes}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
MTIG710imu1kHz
}
\description{
This data set contains six samples collected from MTI-G-710 IMU at 1000 Hz.
}
\examples{
\dontrun{
data(MTIG710imu1kHz)
MTIG710imu1kHz
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
