% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MTIG710imu50Hz}
\alias{MTIG710imu50Hz}
\title{Six samples collected from MTI-G-710 IMU at 50 Hz  over periods of about 6 hours}
\format{An \code{mimu} from object}
\source{
Department of Geomatics Engineering, University of Calgary
}
\usage{
MTIG710imu50Hz
}
\description{
This data set contains five samples collected from MTI-G-710 IMU at 50 Hz.
}
\examples{
\dontrun{
data(MTIG710imu50Hz)
MTIG710imu50Hz
}
}
\author{
Stéphane Guerrier, Gaetan Bakalli
}
\keyword{datasets}
