% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{navchip}
\alias{navchip}
\title{IMU Data over 4 Hours from a navchip sensor}
\format{An \code{imu} with 3,105,250 observations and 6 sensors. The sensors are defined as follows:
\describe{
 \item{\code{Gyro. X}}{Gyroscope X-Axis}
 \item{\code{Gyro. Y}}{Gyroscope Y-Axis}
 \item{\code{Gyro. Z}}{Gyroscope Z-Axis}
 \item{\code{Accel. X}}{Accelerometer X-Axis}
 \item{\code{Accel. Y}}{Accelerometer Y-Axis}
 \item{\code{Accel. Z}}{Accelerometer Z-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
navchip
}
\description{
This data set contains gyroscope and accelerometer data from navchip-based sensors.
}
\examples{
\dontrun{
data(navchip)
navchip
}
}
\author{
JJB
}
\keyword{datasets}

