% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cont.imu1.R
\docType{data}
\name{cont.imu1}
\alias{cont.imu1}
\title{Containment IMU Data from a XSens MTi-G sensor}
\format{A \code{matrix} with 500,000 observations and 1 variable. The variable is as follows:
\describe{
 \item{\code{Gyro. X}}{Gyroscope X-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
cont.imu1
}
\description{
This data set contains gyroscope data
}
\examples{
\dontrun{
data(cont.imu1)
summary(cont.imu1)
}
}
\author{
JJB
}
\keyword{datasets}

