% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imu6.R
\docType{data}
\name{imu6}
\alias{imu6}
\title{IMU Data from a XSens MTi-G sensor}
\format{A \code{matrix} with 873,684 observations and 6 columns. The columns are defined as follows:
\itemize{
 \item{Gyro. X}{Gyroscope X-Axis}
 \item{Gyro. Y}{Gyroscope Y-Axis}
 \item{Gyro. Z}{Gyroscope Z-Axis}
 \item{Accel. X}{Accelerometer X-Axis}
 \item{Accel. Y}{Accelerometer Y-Axis}
 \item{Accel. Z}{Accelerometer Z-Axis}
}}
\source{
Geodetic Engineering Laboratory (TOPO), Swiss Federal Institute of Technology Lausanne (EPFL)
}
\usage{
imu6
}
\description{
This data set contains gyroscope and accelerometer data from a XSens MTi-G sensor.
}
\examples{
\dontrun{
data(imu6)
summary(imu6)
}
}
\author{
JJB
}

