% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imudata.R
\docType{package}
\name{imudata}
\alias{imudata}
\alias{imudata-package}
\title{Inertial Measurement Units (IMU) Data Package}
\description{
This package is meant to serve as a data collection tool for IMU data. 
This data can be used as a means to assess and test methods designed to analyse 
IMU error signals (i.e. long and complex autocorrelated signals). An example method
used for this kind of data is implemented in the GMWM R package which can also model 
the latent models that often characterize this data.
}
\details{
\tabular{ll}{
Package: \tab gmwmdata\cr
Type: \tab Package\cr
Version: \tab 3.0.0\cr
Date: \tab 2018-02-10\cr
License: \tab CC BY-NC-SA 4.0\cr
}
}
\examples{
# Please see the datasets
}
\seealso{
\code{\link[imudata]{cont.imu1}}, \code{\link[imudata]{imu6}}, \code{\link[imudata]{imar.gyro}}, \code{\link[imudata]{ln200.accel}}, \code{\link[imudata]{ln200.gyro}}, \code{\link[imudata]{mtig}}, \code{\link[imudata]{navchip}} and \code{\link[imudata]{kvh}}
}
\author{
Stephane Guerrier \email{stephane@psu.edu},
Roberto Molinari \email{roberto.molinari@hotmail.it},
James Balamuta \email{balamut2@illinois.edu},
Geatan Bakalli \email{Geatan.Bakalli@unige.ch},
Ahmed Radi \email{ahmed.elboraee@ucalgary.ca},
Wenchao Yang \email{wyang40@illinois.edu}

Stephane Guerrier \email{stephane@psu.edu}
}
