% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/imudata.R
\docType{package}
\name{imudata}
\alias{imudata}
\alias{imudata-package}
\title{IMU Data Package for the GMWM Methods Package in R}
\description{
This package is meant to serve as a data collection for IMU data.
The data found here can be used with the GMWM methods R package to model
latent processes within a time series.
}
\details{
\tabular{ll}{
Package: \tab gmwmdata\cr
Type: \tab Package\cr
Version: \tab 0.1.1\cr
Date: \tab 2015-10-14\cr
License: \tab Q + file LICENSE\cr
}
}
\examples{
# Please see the datasets
}
\author{
James Balamuta \email{james.balamuta@gmail.com},
Stephane Guerrier \email{stephane@illinois.edu},
Roberto Molinari \email{roberto.molinari@unige.ch},
Wenchao Yang \email{wyang40@illinois.edu}

James Balamuta \email{james.balamuta@gmail.com}
}

